Release v3.2.0

+++ cookbooks/arcgis-enterprise/recipes/portal.rb
@@ -53,9 +53,10 @@ arcgis_enterprise_portal 'Unpack Portal for ArcGIS' do
   if node['platform'] == 'windows'
     not_if { Utils.product_installed?(node['arcgis']['portal']['product_code']) }
   else
-    not_if { ::File.exist?(::File.join(node['arcgis']['portal']['install_dir'],
-                                       node['arcgis']['portal']['install_subdir'],
-                                       'startportal.sh')) }
+    not_if { EsriProperties.product_installed?(node['arcgis']['run_as_user'],
+                                               node['hostname'],
+                                               node['arcgis']['version'],
+                                               :ArcGISPortal) }
   end
   action :unpack
 end
@@ -70,9 +71,10 @@ arcgis_enterprise_portal 'Install Portal for ArcGIS' do
   if node['platform'] == 'windows'
     not_if { Utils.product_installed?(node['arcgis']['portal']['product_code']) }
   else
-    not_if { ::File.exist?(::File.join(node['arcgis']['portal']['install_dir'],
-                                       node['arcgis']['portal']['install_subdir'],
-                                       'startportal.sh')) }
+    not_if { EsriProperties.product_installed?(node['arcgis']['run_as_user'],
+                                               node['hostname'],
+                                               node['arcgis']['version'],
+                                               :ArcGISPortal) }
   end
   action :install
 end
@@ -94,8 +96,12 @@ arcgis_enterprise_portal 'Authorize Portal for ArcGIS' do
   action :authorize
 end
 
-file node['arcgis']['portal']['cached_authorization_file'] do
-  content File.open(node['arcgis']['portal']['authorization_file'], 'rb') { |file| file.read }
+# Copy portal authorization file to the machine to indicate that portal is already authorized
+file 'Cache portal authorization file' do
+  path node['arcgis']['portal']['cached_authorization_file']
+  if ::File.exists?(node['arcgis']['portal']['authorization_file'])
+    content File.open(node['arcgis']['portal']['authorization_file'], 'rb') { |file| file.read }
+  end
   sensitive true
   subscribes :create, 'arcgis_enterprise_portal[Authorize Portal for ArcGIS]', :immediately
   only_if { node['arcgis']['cache_authorization_files'] }
@@ -103,6 +109,7 @@ file node['arcgis']['portal']['cached_authorization_file'] do
 end
 
 arcgis_enterprise_portal 'Start Portal for ArcGIS' do
+  tomcat_java_opts node['arcgis']['portal']['tomcat_java_opts']
   action :start
 end
 
@@ -127,6 +134,10 @@ arcgis_enterprise_portal 'Create Portal Site' do
   log_level node['arcgis']['portal']['log_level']
   log_dir node['arcgis']['portal']['log_dir']
   max_log_file_age node['arcgis']['portal']['max_log_file_age']
+  upgrade_backup node['arcgis']['portal']['upgrade_backup']
+  upgrade_rollback node['arcgis']['portal']['upgrade_rollback']
+  root_cert node['arcgis']['portal']['root_cert']
+  root_cert_alias node['arcgis']['portal']['root_cert_alias']
   action :create_site
 end
